+++ ProjectRINGSpring/src/main/java/com/ring/bookstore/repository/AccountRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.ring.bookstore.model.Account;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long>{
+public interface AccountRepository extends JpaRepository<Account, Integer>{
 	
 	Optional<Account> findByUserName(String userName);
 }
